+++ km3pipe/math.py
@@ -425,3 +425,85 @@ class loguniform(scipy.stats.rv_continuous):
         # `rvs(size=foo, *args)` does argcheck etc, and sets `self._size`
         return np.power(self._base_of_log, np.random.uniform(
             self._log_low, self._log_high, self._size))
+
+
+def param_names(scipy_dist):
+    """Get names of fit parameters from a ``scipy.rv_*`` distribution."""
+    names = ['loc', 'scale']
+    if scipy_dist.shapes is not None:
+        names += scipy_dist.shapes.split()
+    return names
+
+
+def perc(p=95):
+    """Create symmetric percentiles, with ``p`` coverage."""
+    offset = (100 - p) / 2
+    return offset, 100 - offset
+
+
+def resample_1d(arr):
+    """Resample a 1d array, with replacement."""
+    arr = np.atleast_1d(arr)
+    n = len(arr)
+    idx = np.random.randint(0, n, size=n)
+    return arr[idx]
+
+
+def bootstrap_params(rv_cont, data, n_iter=5):
+    """Bootstrap the fit params of a distribution.
+
+    Parameters
+    ==========
+    rv_cont: scipy.stats.rv_continuous instance
+        The distribution which to fit.
+    data: array-like, 1d
+        The data on which to fit.
+    n_iter: int [default=10]
+        Number of bootstrap iterations.
+    """
+    fit_res = []
+    for i in range(n_iter):
+        params = rv_cont.fit(resample_1d(data))
+        fit_res.append(params)
+    fit_res = np.array(fit_res)
+    return fit_res
+
+
+def param_describe(params, quant=95, axis=0):
+    """Get mean + quantile range from bootstrapped params."""
+    par = np.mean(params, axis=axis)
+    lo, up = perc(quant)
+    p_up = np.percentile(params, up, axis=axis)
+    p_lo = np.percentile(params, lo, axis=axis)
+    return par, p_lo, p_up
+
+
+def bootstrap_fit(rv_cont, data, n_iter=10, quant=95, print_params=True):
+    """Bootstrap a distribution fit + get confidence intervals for the params.
+
+    Parameters
+    ==========
+    rv_cont: scipy.stats.rv_continuous instance
+        The distribution which to fit.
+    data: array-like, 1d
+        The data on which to fit.
+    n_iter: int [default=10]
+        Number of bootstrap iterations.
+    quant: int [default=95]
+        percentile of the confidence limits (default is 95, i.e. 2.5%-97.5%)
+    print_params: bool [default=True]
+        Print a fit summary.
+    """
+    fit_params = bootstrap_params(rv_cont, data, n_iter)
+    par, lo, up = param_describe(fit_params, quant=quant)
+    if not print_params:
+        return par, lo, up
+    names = param_names(rv_cont)
+    maxlen = max([len(s) for s in names])
+    print("--------------")
+    print(rv_cont.name)
+    print("--------------")
+    for i, name in enumerate(names):
+        print("{nam:>{fill}}: {mean:+.3f} ∈ [{lo:+.3f}, {up:+.3f}] ({q}%)".format(
+            nam=name, fill=maxlen, mean=par[i], lo=lo[i], up=up[i], q=quant))
+    return par, lo, up