+++ CHANGELOG.rst
@@ -9,6 +9,8 @@ Unreleased changes
   OID
 * Add ``n_digits`` option in ``kp.io.evt.EvtPump`` for file number index
   when iterating over multiple files.
+* ``kp.math`` has some helpers for bootstrapping confidence intervals
+  when fitting probability distributions via max LLH (in scipy.stats)
 
 Version 7
 ---------